+++ changelog.md
@@ -7,6 +7,7 @@
   * Curse of Obscurity: Blinds you when you take damage (helmet)
   * Curse of Frenzy: Mobs hit gain speed, resistance and strength (chestplate)
   * Curse of Unstable Motion: Gain / lose movement speed at random
+  * Curse of Static Charge: Chance to get struck by lightning
 * Added back enchant glint for cursed only items
 * Durium
   * Sand and Gravel Ores now fall and break